# Improve StringComparer.OrdinalIgnoreCase.Compare perf